+++ src/main/java/com/orejita/games/DAO/Manufacturer/IManufacturerDao.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.orejita.games.Entities.Manufacturer.Manufacturer;
 
 @Repository
-public interface IManufacturerDao extends JpaRepository<Manufacturer, Integer> {
+public interface IManufacturerDao extends JpaRepository<Manufacturer, Long> {
     
 }
